Use better heuristic to search for pages with feed links
In addition to page itself and its top-level domain we now always check
some subdirectories (like '/blog/') if the page was posted under those.
If nothing is found on first pass we remove subdirectories from the end of
URL path and check each new URL until something is found. This is rather
slow, but we have good caching :)